refactor(resources): remove interfaces.go file
